Fixed few bugs(proxy loops). Corrected configuration files. Added spec, README, INSTALL, conf files. Rearranged hierarchy.
XMLRPCService is being used to share memory across the several instances of intelligentmirror. This fix the problem
of looping for the same url while trying to fetch package from remote (because we use the same proxy server to fetch
the package and that ends in loops). More options added to configuration files. Also, README and INSTALL files are
added to for help. Manpage is also added.